Enable heap only on demand
