+++ examples-alloc/alloc_error.rs
