+++ examples-alloc/ble_scanning.rs
