+++ examples-alloc/hardware_test.rs
