+++ examples-alloc/simple_ble.rs
